#23: 회원가입 - 키보드 관련 함수 Extension으로 빼기
