+++ POME/domain/ViewControllers/Register/RegisterViewController.swift
@@ -63,6 +63,7 @@ class RegisterViewController: UIViewController {
         
         maskView.image = Image.photoDefault
         registerView.profileImage.mask = maskView
+        
         // imagePicker delegate
         imagePickerController.delegate = self
         
@@ -71,7 +72,11 @@ class RegisterViewController: UIViewController {
     }
     
     func initButton() {
-        registerView.profileImage.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(self.albumButtonnDidTap)))
+        let albumTapGesture = UITapGestureRecognizer()
+        registerView.profileImage.addGestureRecognizer(albumTapGesture)
+        albumTapGesture.rx.event.bind(onNext: { recognizer in
+            self.albumButtonnDidTap()
+        }).disposed(by: disposeBag)
         registerView.profileButton.rx.tap
             .bind {self.albumButtonnDidTap()}
             .disposed(by: disposeBag)
@@ -99,7 +104,6 @@ class RegisterViewController: UIViewController {
 //            print("name changed : \(newValue) ")
             self.checkValidName(newValue)
         }).disposed(by: disposeBag)
-        
     }
     func checkValidName(_ name: String) {
         if name.count > 0 {
@@ -137,55 +141,4 @@ extension RegisterViewController : UIImagePickerControllerDelegate, UINavigation
         }
         self.dismiss(animated: true, completion: nil)
     }
-}
-// MARK: - TextField & Keyboard Methods
-extension RegisterViewController: UITextFieldDelegate {
-    func addKeyboardNotifications() {
-        // 키보드가 나타날 때 앱에게 알리는 메서드 추가
-        NotificationCenter.default.addObserver(self, selector: #selector(self.keyboardWillAppear(noti:)), name: UIResponder.keyboardWillShowNotification , object: nil)
-        // 키보드가 사라질 때 앱에게 알리는 메서드 추가
-        NotificationCenter.default.addObserver(self, selector: #selector(self.keyboardWillDisappear(noti:)), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    func removeKeyboardNotifications() {
-        // 키보드가 나타날 때 앱에게 알리는 메서드 제거
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification , object: nil)
-        // 키보드가 사라질 때 앱에게 알리는 메서드 제거
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    @objc func keyboardWillAppear(noti: NSNotification) {
-        if let keyboardFrame: NSValue = noti.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue {
-            let keyboardRectangle = keyboardFrame.cgRectValue
-            let keyboardHeight = keyboardRectangle.height + 10
-            let viewHeight = Const.Device.HEIGHT - self.registerView.messageLabel.frame.origin.y
-            print(keyboardHeight, viewHeight)
-            if viewHeight < keyboardHeight {
-                let dif = keyboardHeight - viewHeight
-                self.view.frame.origin.y -= (dif + 20)
-            }
-        }
-        print("keyboard Will appear Execute")
-    }
-    
-    @objc func keyboardWillDisappear(noti: NSNotification) {
-        if self.view.frame.origin.y != restoreFrameValue {
-            if let keyboardFrame: NSValue = noti.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue {
-                let keyboardRectangle = keyboardFrame.cgRectValue
-                let keyboardHeight = keyboardRectangle.height
-                self.view.frame.origin.y += keyboardHeight
-            }
-            print("keyboard Will Disappear Execute")
-        }
-    }
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        print("textFieldShouldReturn Execute")
-        textField.resignFirstResponder()
-        return true
-    }
-
-    func textFieldShouldEndEditing(_ textField: UITextField) -> Bool {
-        print("textFieldShouldEndEditing Execute")
-        self.view.frame.origin.y = self.restoreFrameValue
-        return true
-    }
-    
 }
